refactor move view to user views
