+++ commcare_connect/users/urls.py
@@ -1,6 +1,7 @@
 from django.urls import path
 
 from commcare_connect.users.views import (
+    SMSStatusCallbackView,
     accept_invite,
     create_user_link_view,
     demo_user_tokens,
@@ -19,4 +20,5 @@ urlpatterns = [
     path("start_learn_app/", view=start_learn_app, name="start_learn_app"),
     path("accept_invite/<slug:invite_id>/", view=accept_invite, name="accept_invite"),
     path("demo_users/", view=demo_user_tokens, name="demo_users"),
+    path("sms_status_callback/", SMSStatusCallbackView.as_view(), name="sms_status_callback"),
 ]
