+++ commcare_connect/users/views.py
@@ -4,7 +4,7 @@ from django.contrib.auth.decorators import login_required, user_passes_test
 from django.contrib.auth.mixins import LoginRequiredMixin
 from django.contrib.messages.views import SuccessMessageMixin
 from django.http import HttpResponse
-from django.shortcuts import render
+from django.shortcuts import get_object_or_404, render
 from django.urls import reverse
 from django.utils.decorators import method_decorator
 from django.utils.timezone import now
@@ -15,6 +15,9 @@ from django.views.generic import DetailView, RedirectView, UpdateView, View
 from oauth2_provider.contrib.rest_framework import OAuth2Authentication
 from oauth2_provider.views.mixins import ClientProtectedResourceMixin
 from rest_framework.decorators import api_view, authentication_classes
+from rest_framework.permissions import AllowAny
+from rest_framework.response import Response
+from rest_framework.views import APIView
 
 from commcare_connect.connect_id_client.main import fetch_demo_user_tokens
 from commcare_connect.opportunity.models import Opportunity, OpportunityAccess, UserInvite, UserInviteStatus
@@ -137,3 +140,18 @@ def accept_invite(request, invite_id):
 def demo_user_tokens(request):
     users = fetch_demo_user_tokens()
     return render(request, "users/demo_tokens.html", {"demo_users": users})
+
+
+class SMSStatusCallbackView(APIView):
+    permission_classes = [AllowAny]
+
+    def post(self, *args, **kwargs):
+        message_sid = self.request.data.get("MessageSid", None)
+        message_status = self.request.data.get("MessageStatus", None)
+        user_invite = get_object_or_404(UserInvite, message_sid=message_sid)
+        if message_status == "delivered":
+            user_invite.status = UserInviteStatus.sms_delivered
+        if message_status == "undelivered":
+            user_invite.status = UserInviteStatus.sms_not_delivered
+        user_invite.save()
+        return Response(status=200)
